Activate all features until I can figure out a solution
Code in the master branch is dependent on these variables, and failing
because I didn't plan this properly. I'm turning the feature flags on in
order to provide full fledged functionality as advertised in the README
file.

TODO: Figure out a way to allow people to install only what they want.