Install ~helm-projectile~ for a one-off rare usecase
~helm-projectile~ will bring in ~projectile~, which understands projects
very well and helps solve this usecase for me:

Jumping to a file in the project, when you know the name of the file but
it is not in your ~ido-virtual-buffers~ or ~recentf~ list. This is a
super rare condition, but I found myself needing it when I was recently
reading the code of an old project that I hadn't worked on it ages.