Require org-tree-slide explicitly
For some weird reason, this needs to be done to avoid void-variable
errors. I have a feeling it has to do with `:required nil` in the
recipe, but since el-get does not support such a keyword in the
documentation, I'm not sure. For the moment, simply adding an
explicit require.