Install lispy
Note that this commit depends on the latest dev branch of my el-get
repo, for the lispy and zoutline recipes.

Also, I have created a personal recipe for swiper since the makeinfo
command is failing on my Mac. I will look into this later to
install the missing software packages on the Mac.

Note that `lispy-mode` is installed alongside `pardedit-mode` via the
`turn-on-paredit` function, since we already maintain a list of lispy
modes