Re-arrange utf-8 settings
This was an attempt to fix Devanagari rendering on my machine,
following the Mastering Emacs blog. But it did not work.

What eventually worked was installing the correct Devanagari font on
my Mac OS. I cannot explain why this was working earlier, broke, and
then was fixed by installing a font. But it was.