Move all the prettification under my pretty flag
Emacs is running slow / lagging badly currently. There are two
culprits:

1. All the pretty symbols : This is not a problem actually, but I've
been looking for excuses to move it under the flag and go back to
austere Emacs, so this is it.

2. The `gcmh` package : This hacks memory management and garbage
collection in Emacs, and I don't think it works very well with the
latest version of Emacs.

These changes have made Emacs super snappy again! Woohoo! 🙌