Do not generate inline CSS when exporting from org-mode
The change to `org-html-htmlize` vars ensures that HTML is generated
with the appropriate classes, but the styling is left to an external
stylesheet. This lets me add
```org
    #+HTML_HEAD: <link rel="stylesheet" type="text/css" href="https://gongzhitaao.org/orgcss/org.css"/>
```
to my org files and use https://github.com/gongzhitaao/orgcss to style
my exported data.

Also, when exporting to PDF, use Emacs, `pdf-tools` to open it up for
viewing.