Add a journaling yasnippet
I am using the `denote-template` system to enter a key to a
`yasnippet`, which then expands into a full template. This way I can
leverage the power of a full templating system when creating notes
with `denote`.

Even better would have been to tap into `org-capture`, which, in my
opinion, is an even more powerful templating system just for
`org-mode`. But we take what we get.