Install denote
Experimenting with Denote to see if I should move away from org-brain.
Basically, this is an experiment in whether one-file-per-note is
better than my existing one-file-per-area approach.

I like the denote code-base, it's easy to read and understand. At the
moment, I'm not convinced it's better than org-brain. But that's why
the experiment.